agents/requirements.txt
fastapi==0.115.12
httpx==0.28.1
python-dotenv==1.0.1
python-telegram-bot==21.11.1
uvicorn==0.34.2
openai>=1.30.0
python-docx>=1.1.0